refactor(ai-agent): split openclaw/agent.rs into submodules (Stage 9.3, M5)
Purely mechanical split — function bodies byte-identical.

- openclaw/agent/mod.rs:            struct + constructor + IAgentTask impl + event relay
- openclaw/agent/confirmations.rs:  confirm / get_confirmations / check_approval helpers
- openclaw/agent/spawn_helpers.rs:  build_spawn_config + gateway-ready helpers + tests

Public path `crate::manager::openclaw::agent::OpenClawAgentManager` unchanged.
Each sub-file ≤ 400 lines. No behavior change.
